refactor: split skill profile step
